pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "influx-agent"
version = "0.1.0"
description = "Thin HTTP → LLM proxy wrapping Ollama with an OpenAI-compatible API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic-settings>=2.2.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-ollama>=0.1.0",
    "influxdb==5.3.1",
    "pandas>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
